# backend/swagger.json
{
  "openapi": "3.0.3",
  "paths": {
...
    "/beta-registration/public": {
      "post": {
        "description": "Register a wallet address for public beta with Twitter verification",
        "operationId": "BetaRegistrationController_registerPublic",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Public registration request with Twitter details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePublicRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetaRegistrationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input data"
          },
          "409": {
            "description": "Conflict - wallet already registered"
          }
        },
        "summary": "Register for public beta",
        "tags": [
          "Beta Registration"
        ]
      }
    },
    "/beta-registration/community": {
      "post": {
        "description": "Register a wallet address for community beta",
        "operationId": "BetaRegistrationController_registerCommunity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Community registration request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommunityRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetaRegistrationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input data or community limit reached"
          },
          "404": {
            "description": "Community not found"
          },
          "409": {
            "description": "Conflict - wallet already registered"
          }
        },
        "summary": "Register for community beta",
        "tags": [
          "Beta Registration"
        ]
      }
    },
...
    "/beta-registration/community/{communityCode}/stats": {
      "get": {
        "description": "Get registration statistics for a specific community",
        "operationId": "BetaRegistrationController_getCommunityStats",
        "parameters": [
          {
            "name": "communityCode",
            "required": true,
            "in": "path",
            "description": "Community code",
            "schema": {
              "example": "CRYPTO_BROS",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Community statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityStatsDto"
                }
              }
            }
          },
          "404": {
            "description": "Community not found"
          }
        },
        "summary": "Get community registration statistics",
        "tags": [
          "Beta Registration"
        ]
      }
    },
    "/beta-registration/community/create": {
      "post": {
        "description": "Create a new beta community with registration limits",
        "operationId": "BetaRegistrationController_createCommunity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Community creation request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommunityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Community created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input data"
          },
          "409": {
            "description": "Conflict - community code already exists"
          }
        },
        "summary": "Create a new community",
        "tags": [
          "Beta Registration"
        ]
      }
    },
...
  },
  "info": {
    "title": "Brother Terminal Trading API",
    "description": "API for managing trading positions and orders on the Brother Terminal platform for Hyperliquid Engine.",
    "version": "3.0.0",
    "contact": {
      "name": "Brother Terminal",
      "url": "https://brother-terminal.xyz",
      "email": "dev@brother-terminal.xyz"
    }
  },
...
  "servers": [
    {
      "url": "https://api.brother-terminal.xyz",
      "description": "Production (Mainnet)"
    },
    {
      "url": "https://api-testnet.brother-terminal.xyz",
      "description": "Testnet"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePublicRegistrationDto": {
        "type": "object",
        "properties": {
          "walletAddress": {
            "type": "string",
            "description": "User wallet address",
            "example": "0x1234567890123456789012345678901234567890",
            "pattern": "^0x[a-fA-F0-9]{40}$"
          },
          "twitterHandle": {
            "type": "string",
            "description": "Twitter handle of the user",
            "example": "@john_doe"
          },
          "tweetUrl": {
            "type": "string",
            "description": "URL of the tweet",
            "example": "https://twitter.com/user/status/123456789"
          }
        },
        "required": [
          "walletAddress",
          "twitterHandle",
          "tweetUrl"
        ]
      },
      "BetaRegistrationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Registration ID",
            "example": "reg_123456789"
          },
          "walletAddress": {
            "type": "string",
            "description": "Wallet address",
            "example": "0x1234567890123456789012345678901234567890"
          },
          "registrationType": {
            "type": "string",
            "description": "Registration type",
            "enum": [
              "PUBLIC",
              "COMMUNITY"
            ],
            "example": "PUBLIC"
          },
          "communityCode": {
            "type": "string",
            "description": "Community code (only for community registrations)",
            "example": "CRYPTO_BROS"
          },
          "twitterHandle": {
            "type": "string",
            "description": "Twitter handle (only for public registrations)",
            "example": "@john_doe"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Registration creation date",
            "example": "2025-01-15T10:30:00Z"
          }
        },
        "required": [
          "id",
          "walletAddress",
          "registrationType",
          "createdAt"
        ]
      },
      "CreateCommunityRegistrationDto": {
        "type": "object",
        "properties": {
          "walletAddress": {
            "type": "string",
            "description": "User wallet address",
            "example": "0x1234567890123456789012345678901234567890",
            "pattern": "^0x[a-fA-F0-9]{40}$"
          },
          "communityCode": {
            "type": "string",
            "description": "Community code",
            "example": "CRYPTO_BROS"
          }
        },
        "required": [
          "walletAddress",
          "communityCode"
        ]
      },
...
      "CommunityStatsDto": {
        "type": "object",
        "properties": {
          "communityCode": {
            "type": "string",
            "description": "Community code",
            "example": "CRYPTO_BROS"
          },
          "name": {
            "type": "string",
            "description": "Community name",
            "example": "Crypto Brothers Community"
          },
          "registrationCount": {
            "type": "number",
            "description": "Community registration count",
            "example": 45
          },
          "registrationLimit": {
            "type": "number",
            "description": "Community registration limit",
            "example": 100
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether community is active",
            "example": true
          }
        },
        "required": [
          "communityCode",
          "name",
          "registrationCount",
          "registrationLimit",
          "isActive"
        ]
      },
      "CreateCommunityDto": {
        "type": "object",
        "properties": {
          "communityCode": {
            "type": "string",
            "description": "Community code (unique identifier)",
            "example": "CRYPTO_BROS"
          },
          "name": {
            "type": "string",
            "description": "Community name",
            "example": "Crypto Brothers Community"
          },
          "description": {
            "type": "string",
            "description": "Community description",
            "example": "A community of crypto enthusiasts"
          },
          "registrationLimit": {
            "type": "number",
            "description": "Registration limit for this community",
            "example": 100,
            "default": 1000
          }
        },
        "required": [
          "communityCode",
          "name",
          "registrationLimit"
        ]
      },
      "CommunityResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Community ID",
            "example": "com_123456789"
          },
          "communityCode": {
            "type": "string",
            "description": "Community code",
            "example": "CRYPTO_BROS"
          },
          "name": {
            "type": "string",
            "description": "Community name",
            "example": "Crypto Brothers Community"
          },
          "description": {
            "type": "string",
            "description": "Community description",
            "example": "A community of crypto enthusiasts"
          },
          "registrationLimit": {
            "type": "number",
            "description": "Registration limit",
            "example": 100
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether community is active",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Community creation date",
            "example": "2025-01-15T10:30:00Z"
          }
        },
        "required": [
          "id",
          "communityCode",
          "name",
          "description",
          "registrationLimit",
          "isActive",
          "createdAt"
        ]
      },
...
    }
  }
}